SCRD-2385 Fix apache site management on SUSE
Without this patch, the playbooks implement start and stop tasks for
Apache sites that depend on the a2ensite and a2dissite functionality,
with the intention of being able to enable and disable particular
services without stopping and starting Apache itself. We previously
naively ensured that the vhost would be configured on SUSE but neglected
to ensure the start and stop playbooks worked. In order to replicate
this functionality without using the a2ensite and a2dissite utilities,
the playbooks must instead add and remove the vhost file from the main
config directory. This patch also removes an unnecessary site enablement
task from the start playbook since it is repeated later in the playbook.

(Update copyrights)

Change-Id: Id00ff8d02517a1fbaf4ba00edf8c20d6b98a3a95