Secure the home and shell for ceilometer (bsc#1083602)
The Ansible default behavior for creating a user is not very
secure. As ceilometer doesn't need a shell, specify
/sbin/nologin shell.
Unfortunately, the mod_wsgi does need a home directory, so
create a home dir under /var/lib/ceilometer.

Note this is done for both legacy and suseified installs, due
to the behavior of install_package module not installing the
rpm as you might expect. Eventually user creation in the Ardana
scripts can be dropped and just let the user creation from
the .rpm stand, but that will take some additional work.

Change-Id: I7a81f3eda405127d8c970b8cf09ef637ad5c008d