Don't change the ceilometer user if it is in use (bsc#1092182)
It was found in update testing that the ansible script was attempting
to change the ceilometer user but there are processes running as
the ceilometer user, which caused a failure.

This change catches the error from the attempt to change the existing
user and if the failure is due to existing processes under the
ceilometer user it continues on (and does not attempt to create
a home dir).  The advantage of this approach is that if the ceilometer
processes are stopped before the install.yml is run then the change
can be made.

Change-Id: I6f9e77852ef28b57a9627ac2432e94c73fea6b23